\name{grid-internal}
\alias{.get.genes}
\alias{.classify_sample}
\alias{.getmarker_ps}
\alias{.collapseRows}
\alias{.get_mpsmat}
\title{Internal sampleClassifier Functions}
\description{
 Internal sampleClassifier functions
}
\details{
These are not intended to be called by the user.
}
\keyword{ internal }
